+++ src/exercicio.ts
@@ -1,0 +1,5 @@
+const multiplicar = (num1: number, num2: number): number => { return num1 * num2 }
+console.log(multiplicar(5, 2))
+
+const saudacao = (nome: string): string => { return `Olá ${nome}` }
+console.log(saudacao('Bruno Daniel'))